pyproject.toml
[tool.poetry]
name = "amnesty-tracker"
version = "0.1.0"
description = ""
authors = ["licht1stein <mb@blaster.ai>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
requests = "^2.26.0"
pendulum = "^2.1.2"
envparse = "^0.2.0"
python-telegram-bot = "^13.7"
html5lib = "^1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
